requirements.txt
annotated-types==0.7.0
anyio==4.9.0
assemblyai==0.28.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.9
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0